Move store to /src and remove local functions from store
